fix warnings for windows

+++ MaeParser.cpp
@@ -21,12 +21,12 @@ static bool property_key_author_name(Buffer& buffer, char*& save);
 
 static std::string outer_block_name(Buffer& buffer);
 
-void read_exception::format(int line_number, int column, const char* msg)
+void read_exception::format(size_t line_number, size_t column, const char* msg)
 {
 #ifdef _MSC_VER
-    _snprintf(m_msg, MAEPARSER_EXCEPTION_BUFFER_SIZE, "Line %d, column %d: %s\n", line_number,
+    _snprintf(m_msg, MAEPARSER_EXCEPTION_BUFFER_SIZE, "Line %zu, column %zu: %s\n", line_number,
 #else
-    snprintf(m_msg, MAEPARSER_EXCEPTION_BUFFER_SIZE, "Line %d, column %d: %s\n", line_number,
+    snprintf(m_msg, MAEPARSER_EXCEPTION_BUFFER_SIZE, "Line %zu, column %zu: %s\n", line_number,
 #endif
              column, msg);
     m_msg[MAEPARSER_EXCEPTION_BUFFER_SIZE - 1] = '\0';
