+++ MaeParser.hpp
@@ -74,7 +74,7 @@ class EXPORT_MAEPARSER read_exception : public std::exception
   private:
     char m_msg[MAEPARSER_EXCEPTION_BUFFER_SIZE];
 
-    void format(int line_number, int column, const char* msg);
+    void format(size_t line_number, size_t column, const char* msg);
 
   public:
     read_exception(const Buffer& buffer, const char* msg)
@@ -82,7 +82,7 @@ class EXPORT_MAEPARSER read_exception : public std::exception
         format(buffer.line_number, buffer.getColumn(), msg);
     }
 
-    read_exception(int line_number, int column, const char* msg)
+    read_exception(size_t line_number, size_t column, const char* msg)
     {
         format(line_number, column, msg);
     }
